+++ packages/angular_devkit/build_angular/src/utils/action-executor.ts
@@ -7,9 +7,9 @@
  */
 
 import Piscina from 'piscina';
+import { InlineOptions } from './bundle-inline-options';
 import { maxWorkers } from './environment-options';
 import { I18nOptions } from './i18n-options';
-import { InlineOptions } from './process-bundle';
 
 const workerFile = require.resolve('./process-bundle');
