+++ packages/angular_devkit/build_angular/src/utils/i18n-inlining.ts
@@ -11,10 +11,10 @@ import { EmittedFiles } from '@angular-devkit/build-webpack';
 import * as fs from 'fs';
 import * as path from 'path';
 import { BundleActionExecutor } from './action-executor';
+import { InlineOptions } from './bundle-inline-options';
 import { copyAssets } from './copy-assets';
 import { assertIsError } from './error';
 import { I18nOptions } from './i18n-options';
-import { InlineOptions } from './process-bundle';
 import { Spinner } from './spinner';
 
 function emittedFilesToInlineOptions(
